🚚 style: JSONError 타입 파일 분리 #6
