+++ Expo1900/Expo1900/Model/JSONParser.swift
@@ -1,16 +1,5 @@
 import UIKit
 
-enum JSONError: LocalizedError {
-    case notFound
-    
-    var errorDescription: String? {
-        switch self {
-        case .notFound:
-            return "데이터를 불러오지 못했습니다."
-        }
-    }
-}
-
 enum JSONParser<T: Decodable> {
     static func decode(fileName: String) -> Result<T, JSONError>  {
         guard let dataAsset: NSDataAsset = NSDataAsset(name: fileName) else {
